+++ src/main/java/com/manchui/domain/image/repository/ImageRepository.java
@@ -1,8 +1,9 @@
-package com.manchui.domain.repository;
+package com.manchui.domain.image.repository;
 
-import com.manchui.domain.entity.Image;
+import com.manchui.domain.image.entity.Image;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ImageRepository extends JpaRepository<Image, Long> {
     Image findByGatheringId(Long gatheringId);
+
 }
