+++ src/main/java/com/manchui/domain/image/service/S3Uploader.java
@@ -1,4 +1,4 @@
-package com.manchui.domain.service;
+package com.manchui.domain.image.service;
 
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.CannedAccessControlList;
